add conda env and switch to snakemake directory structure
