+++ .flake8
@@ -5,3 +5,4 @@ max-line-length = 81
 select = A,C,E,F,W,B,B950
 #B305 doesn't like `.next()` that is a key Tree method.
 ignore = E203, E501, W503, B305
+builtins = snakemake
