+++ workflow/scripts/rule_full_simplify.py
@@ -1,0 +1,5 @@
+import tskit
+
+ts = tskit.load(snakemake.input[0])
+ts = ts.simplify()
+ts.dump(snakemake.output[0])
