+++ workflow/tskit_reqs.yaml
@@ -1,0 +1,20 @@
+channels:
+  - conda-forge
+  - bioconda
+dependencies:
+  - tsinfer=0.4.1
+  - tskit=0.6.3
+  - sgkit=0.10.0
+  - aiohttp=3.11.18
+  - cyvcf2=0.31.1
+  - requests=2.32.3
+  - yarl=1.20.0
+  - msprime=1.3.4
+  - pysam=0.23.0
+  - bokeh=3.7.3
+  - pyfaidx=0.8.1.4
+  - matplotlib=3.10.3
+  - bio2zarr=0.1.5
+  - pip
+  - pip:
+      - tsdate==0.2.2
